Most tests run

+++ tests/compatibility_tests/base_test_case/base_compatibility_test.py
@@ -2,7 +2,7 @@ import json
 import logging
 from abc import abstractmethod, ABC
 from pathlib import Path
-from marqo_test import MarqoTestCase
+from tests.compatibility_tests.base_test_case.marqo_test import MarqoTestCase
 
 
 class BaseCompatibilityTestCase(MarqoTestCase, ABC):
@@ -63,8 +63,11 @@ class BaseCompatibilityTestCase(MarqoTestCase, ABC):
     def delete_file(cls):
         """Delete the results file."""
         filepath = cls.get_results_file_path()
-        filepath.unlink()
-        cls.logger.debug(f"Results file deleted: {filepath}")
+        if filepath.exists():
+            filepath.unlink()
+            cls.logger.debug(f"Results file deleted: {filepath}")
+        else:
+            cls.logger.debug(f"Not deleting, as the results file was never created in the first place.")
 
     @abstractmethod
     def prepare(self):
